requirements.txt
python-dotenv
themoviedb
pyyaml
